feat(backend): complete backend functionalities, product management, user authntication and authorization & error handling

+++ backend/src/services/product.service.ts
@@ -1,6 +1,17 @@
 import prisma from "../config/database";
 import { Prisma } from "@prisma/client";
+// import cloudinary from "../config/cloudinary";
 
+
+/**
+ * Function to add a product
+ * @param name 
+ * @param description 
+ * @param price 
+ * @param stockQuantity 
+ * @param categoryId 
+ * @returns 
+ */
 const addProduct = async (name: string, description: string, price: number, image: string, stockQuantity: number, categoryId: number) => {
   const newProduct = await prisma.product.create({
     data: {
@@ -14,6 +25,61 @@ const addProduct = async (name: string, description: string, price: number, imag
   });
   return newProduct;
 };
+
+// /**
+//  * Function to add a product with image upload
+//  * @param name 
+//  * @param description 
+//  * @param price 
+//  * @param stockQuantity 
+//  * @param categoryId 
+//  * @param imagePath 
+//  * @returns 
+//  */
+// const addProduct = async (name: string, description: string, price: number, stockQuantity: number, categoryId: string, imagePath: string) => {
+//   try {
+//     // Convert categoryId to integer
+//     const categoryIntId = parseInt(categoryId, 10);
+
+//     // Ensure the category exists
+//     const category = await prisma.category.findUnique({ where: { id: categoryIntId } });
+//     if (!category) {
+//       throw new Error("Category not found");
+//     }
+
+//     console.log(`Uploading image: ${imagePath}`);
+
+//     // Upload image to Cloudinary
+//     const result = await cloudinary.uploader.upload(imagePath);
+//     console.log(`Upload result: ${JSON.stringify(result)}`);
+
+//     if (!result || !result.secure_url) {
+//       throw new Error("Invalid image file");
+//     }
+
+//     // Create the product
+//     const newProduct = await prisma.product.create({
+//       data: {
+//         name,
+//         description,
+//         price,
+//         image: result.secure_url,
+//         stockQuantity,
+//         categoryId: categoryIntId
+//       }
+//     });
+
+//     console.log(`Product created: ${JSON.stringify(newProduct)}`);
+//     return newProduct;
+//   } catch (error) {
+//     console.error(error);
+//     if (error instanceof Error) {
+//       throw new Error(error.message);
+//     } else {
+//       throw new Error("An unknown error occurred");
+//     }
+//   }
+// };
 
 const getAllProducts = async () => {
   return await prisma.product.findMany();
@@ -31,6 +97,17 @@ const updateProduct = async (id: number, data: Prisma.ProductUpdateInput) => {
 };
 
 const deleteProduct = async (id: number) => {
+  /**
+   * Check if the product exists
+   */
+  const product = await prisma.product.findUnique({ where: { id } });
+  if (!product) {
+    throw new Error("Product not found");
+  }
+
+  /**
+   * Delete the product
+   */
   await prisma.product.delete({ where: { id } });
 };
